Restructure src directory
