+++ src/capiThriftModels.ts
