+++ src/components/news/article.tsx
@@ -7,14 +7,14 @@ import ArticleStandfirst from './articleStandfirst';
 import ArticleByline from './articleByline';
 import ArticleBody from './articleBody';
 import Tags from 'components/shared/tags';
-import { Content } from 'types/capi-thrift-models';
+import { Content } from 'capiThriftModels';
 import { darkModeCss, articleWidthStyles } from 'styles';
 import { palette } from '@guardian/src-foundations';
 import { from, breakpoints } from '@guardian/src-foundations/mq';
 import { css } from '@emotion/core';
 import { Keyline } from 'components/shared/keyline';
-import { isFeature, articleSeries, articleContributors, articleMainImage } from 'types/capi';
-import { getPillarStyles, pillarFromString } from 'types/pillar';
+import { isFeature, articleSeries, articleContributors, articleMainImage } from 'capi';
+import { getPillarStyles, pillarFromString } from 'pillar';
 
 export interface ArticleProps {
     capi: Content;
