--- src/html/liveblogTemplate.html
@@ -1,78 +1,0 @@
-<!DOCTYPE html>
-<html lang="en-US">
-<head>
-    <title>Liveblog</title>
-    <meta charset="UTF-8" />
-    <meta id="twitter-theme" name="twitter:widgets:theme" content="light" />
-    <meta name="viewport" content="initial-scale=1, maximum-scale=1" />
-    <style>
-      @font-face {
-        font-family: "Guardian Egyptian Web";
-        src: url('/public/fonts/egyptiantext-regular.otf');
-      }
-
-      @font-face {
-        font-family: "Guardian Egyptian Web";
-        font-weight: 500;
-        src: url("/public/fonts/egyptiantext-medium.otf");
-      }
-
-      @font-face {
-        font-family: "Guardian Egyptian Web";
-        font-weight: bold;
-        src: url("/public/fonts/egyptiantext-medium.otf");
-      }
-
-      @font-face {
-        font-family: "Guardian Text Sans Web";
-        src: url('/public/fonts/GuardianSansWeb-Regular.ttf');
-      }
-
-      @font-face {
-        font-family: "Guardian Headline";
-        font-weight: 200;
-        src: url("/public/fonts/GHGuardianHeadline-Light.ttf");
-      }
-
-      @font-face {
-        font-family: "Guardian Headline";
-        font-weight: 400;
-        src: url("/public/fonts/GHGuardianHeadline-Regular.ttf");
-      }
-
-      @font-face {
-        font-family: "Guardian Headline";
-        font-weight: 500;
-        src: url("/public/fonts/GHGuardianHeadline-Medium.ttf");
-      }
-
-      @font-face {
-        font-family: "Guardian Headline";
-        font-weight: 700;
-        src: url("/public/fonts/GHGuardianHeadline-Bold.ttf");
-      }
-
-      @font-face {
-        font-family: "Guardian Icons";
-        src: url('/public/fonts/icons.otf');
-      }
-
-      html {
-          font-size: 62.5%;
-      }
-
-      body {
-        margin: 0;
-        font-family: 'Guardian Egyptian Web';
-        font-size: 1.6em;
-        line-height: 1.5em;
-        overflow-x: hidden;
-      }
-    </style>
-</head>
-<body>
-  <div id="root"></div>
-  <script src="https://platform.twitter.com/widgets.js"></script>
-  <script src="/liveblog.js"></script>
-</body>
-</html>
