+++ src/renderBlocks.ts
@@ -6,8 +6,8 @@ import jsdom from 'jsdom';
 import { Result, Err, fromUnsafe } from './types/result';
 import { imageBlock } from './components/blocks/image';
 import { insertAdPlaceholders } from './ads';
-import { transform } from 'utils/contentTransformations';
-import { BlockElement } from 'types/capi-thrift-models';
+import { transform } from 'contentTransformations';
+import { BlockElement } from 'capiThriftModels';
 
 
 // ----- Setup ----- //
