+++ src/server/ssmConfig.ts
@@ -1,5 +1,5 @@
-import {ssm} from "./aws";
-import {App, Stack, Stage} from "./appIdentity";
+import { ssm } from './aws';
+import { App, Stack, Stage } from './appIdentity';
 import { Option, Some, None } from 'types/option';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
